+++ docs/GETTING-STARTED-MANUALLY.md
@@ -31,14 +31,14 @@ $EDITOR .env
 **Never check in your `.env` file!**
 It contains your most private information.
 
-**This project can be used as a Docker container!** Use the `docker run` commands below, after building the container:
+**This project can be used as a Docker container!** Use the `docker init` commands below, after building the container:
 `docker build -t jsappme/node-binance-trader .`
 
 **To start the server** to save pair data, define strategies and emit trading signals:
 ```
-npm run start
+npm init start
 // or
-docker run -d --name node-binance-trader -v "$PWD/.env:/srv/app/.env" -p 4000:4000 jsappme/node-binance-trader npm run start
+docker init -d --name node-binance-trader -v "$PWD/.env:/srv/app/.env" -p 4000:4000 jsappme/node-binance-trader npm init start
 ```
 
 **To start the auto trader** to monitor strategies and signals received from the server or the NBT Hub:
@@ -46,14 +46,14 @@ docker run -d --name node-binance-trader -v "$PWD/.env:/srv/app/.env" -p 4000:40
 <i>Important note: Always make sure to have some BNB available on your corresponding wallet to pay for the fees.</i>
 
 ```
-npm run trader
+npm init trader
 // or
-docker run -d --name node-binance-trader -v "$PWD/.env:/srv/app/.env" jsappme/node-binance-trader npm run trader
+docker init -d --name node-binance-trader -v "$PWD/.env:/srv/app/.env" jsappme/node-binance-trader npm init trader
 ```
 
 **To backtest** strategies using the data recorded by the server:
 ```
-npm run backtest
+npm init backtest
 // or
-docker run -d --name node-binance-trader -v "$PWD/.env:/srv/app/.env" jsappme/node-binance-trader npm run backtest
+docker init -d --name node-binance-trader -v "$PWD/.env:/srv/app/.env" jsappme/node-binance-trader npm init backtest
 ```
